Add files via upload
There is a significant difference between a website and a landing page.
A landing page consists of just one page, while a website may contain multiple pages.
Your landing page should include all important information in a concise and clear manner.
It should effectively prompt users to take a call to action.
It should be user-friendly with a simple and attractive design.
In addition to this, I'd like to mention that as software developers, our main job is to solve specific problems through our products. While browsing for internships online, I noticed that there was no direct way to discover and apply for internships. So, I decided to solve this problem with my product, a landing page. I present to you the one-stop solution for every tech enthusiast interested in gaining internships online: CodSoft Landing Page.

As stated in the document, it is purely made with HTML and Vanilla CSS, with just one Bootstrap library used.
It includes: 
Aim of the website: 
To provide students with internship opportunities.

Description: Includes a small description about the the different domains in which student can apply for the respective internship.

Internship certificate: 
It includes an attached sample of the certificate that students will gain after successful completion of their internship.

Form : 
It includes a form in which students can directly apply for the internships.

Contact us : 
It includes various ways in which the students can get in touch with Codsoft.

Technology used: HTML, Vanilla CSS, and Bootstrap.
Teacher: Shradha Khapra 
Mentor: AMAN DHATTARWAL 
Intern at: CodSoft 

I have unlocked the door of creativity."